# lib/units/api/swagger/api_v1_generated.json
{
  "swagger": "2.0",
  "info": {
    "version": "2.3.0",
    "title": "Smartphone Test Farm",
    "description": "Control and manages real Smartphone devices from browser and restful apis",
    "license": {
      "name": "Apache-2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "contact": {
      "name": "STF Team",
      "email": "contact@openstf.io",
      "url": "http://openstf.io/"
    }
  },
  "basePath": "/api/v1",
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json",
    "application/octet-stream"
  ],
  "produces": [
    "application/json"
  ],
...
  "paths": {
...
    "/projects/getprojects": {
      "get": {
        "summary": "Get Projects",
        "description": "Get All Projects with conditon",
        "operationId": "getProjects",
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "name": "requirement",
            "in": "query",
            "description": "Projects requirement",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Get Projects information",
            "schema": {
              "$ref": "#/definitions/ProjectsResponse"
            }
          },
          "default": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "accessTokenAuth": []
          }
        ]
      }
    },
    "/projects/updateProjects": {
      "get": {
        "summary": "updateProjects",
        "description": "update All Projects ",
        "operationId": "updateProjects",
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "user",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Update Projects Syn",
            "schema": {
              "$ref": "#/definitions/UpdateProjects"
            }
          },
          "default": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "accessTokenAuth": []
          }
        ]
      }
    },
    "/projects/getProjectSynBar": {
      "get": {
        "summary": "getProjectSynBar",
        "description": "get  getProjectSynBar ",
        "operationId": "getProjectSynBar",
        "tags": [
          "projects"
        ],
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "user",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "get getProjectSynBar",
            "schema": {
              "$ref": "#/definitions/GetProjectSynBar"
            }
          },
          "default": {
            "description": "Unexpected Error",
            "schema": {
              "$ref": "#/definitions/ErrorResponse"
            }
          }
        },
        "security": [
          {
            "accessTokenAuth": []
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "ErrorResponse": {
      "required": [
        "message"
      ],
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "accessTokenAuth": {
      "type": "apiKey",
      "name": "authorization",
      "in": "header"
    }
  }
}